refactor: move event-inputs to own file
