+++ apps/ensapi/src/omnigraph-api/schema/domain.ts
@@ -38,7 +38,8 @@ import {
   DomainsOrderInput,
   SubdomainsWhereInput,
 } from "@/omnigraph-api/schema/domain-inputs";
-import { EventRef, EventsWhereInput } from "@/omnigraph-api/schema/event";
+import { EventRef } from "@/omnigraph-api/schema/event";
+import { EventsWhereInput } from "@/omnigraph-api/schema/event-inputs";
 import { LabelRef } from "@/omnigraph-api/schema/label";
 import { PermissionsUserRef } from "@/omnigraph-api/schema/permissions";
 import { RegistrationInterfaceRef } from "@/omnigraph-api/schema/registration";
@@ -319,12 +320,13 @@ ENSv2DomainRef.implement({
         where: t.arg({ type: DomainPermissionsWhereInput }),
       },
       resolve: (parent, args) => {
-        // NOTE: avoid inArray([]) runtime error by short-circuit to an explicit empty result
         const userScope = (() => {
           const user = args.where?.user;
           if (!user) return undefined;
 
           const userIn = user.in ?? [user.eq];
+
+          // NOTE: avoid inArray([]) runtime error by short-circuit to an explicit empty result
           if (userIn.length === 0) return sql`false`;
 
           return inArray(ensIndexerSchema.permissionsUser.user, userIn);
