+++ apps/ensapi/src/omnigraph-api/schema/resolver.ts
@@ -19,7 +19,8 @@ import { getModelId } from "@/omnigraph-api/lib/get-model-id";
 import { lazyConnection } from "@/omnigraph-api/lib/lazy-connection";
 import { AccountIdInput, AccountIdRef } from "@/omnigraph-api/schema/account-id";
 import { ID_PAGINATED_CONNECTION_ARGS } from "@/omnigraph-api/schema/constants";
-import { EventRef, EventsWhereInput } from "@/omnigraph-api/schema/event";
+import { EventRef } from "@/omnigraph-api/schema/event";
+import { EventsWhereInput } from "@/omnigraph-api/schema/event-inputs";
 import { NameOrNodeInput } from "@/omnigraph-api/schema/name-or-node";
 import { PermissionsRef } from "@/omnigraph-api/schema/permissions";
 import { RegistryInterfaceRef } from "@/omnigraph-api/schema/registry";
